% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluate.R
\name{Evaluate}
\alias{Evaluate}
\title{Module for evaluation}
\usage{
Evaluate(envir = .GlobalEnv, quiet = FALSE)
}
\arguments{
\item{envir}{(environment)}

\item{quiet}{(logical)}
}
\description{
Module for evaluation
}
