% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layout.R
\name{Layout}
\alias{Layout}
\title{Layout Module}
\usage{
Layout(border = "*", padding = 1)
}
\arguments{
\item{border}{(character) see \link{newPresentation}.}

\item{padding}{(integer) see \link{newPresentation}.}
}
\description{
The function returns a list with it's arguments as elements. You do not need
to use this function interactively: see \link{newPresentation}.
}
