% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Presentation.R
\name{Presentation}
\alias{Presentation}
\title{Create a presentation}
\usage{
Presentation(fileName, read, render, slideDeck, evaluate)
}
\arguments{
\item{fileName}{(fileName)}

\item{read}{(Read)}

\item{render}{(Render)}

\item{slideDeck}{(SlideDeck)}
}
\description{
Create a presentation
}
