% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.R
\name{Read}
\alias{Read}
\title{Read Module}
\usage{
Read()
}
\description{
Read creates a new module with one method: 'auto'. This method selects an
appropriate procedure for reading in the content of the file depending on the
file extension. It should not be necessary to call this function
interactively, please go to the docs for \link{newPresentation}.
}
\seealso{
\link{newPresentation}
}
