% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Render.R
\name{Render}
\alias{Render}
\title{Render Module}
\usage{
Render(layout, screen)
}
\arguments{
\item{layout}{(Layout) an instance of \link{Layout}}

\item{screen}{(Screen) an instance of \link{Screen}}
}
\description{
This module should not be used interactively. See \link{newPresentation}
instead. The module handles everything around printing a slide to the
console.
}
