% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Screen.R
\name{Screen}
\alias{Screen}
\title{Screen Module}
\usage{
Screen(.width = NULL, .height = NULL)
}
\arguments{
\item{.width, .height}{(NULL | integer) see \link{newPresentation}.}
}
\description{
A list with two functions for determining the height and width of the space
in which to print the slide. The two methods do not have any arguments. You do not need to use this function interactively: see \link{newPresentation}.
}
