% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slide.R
\name{Slide}
\alias{Slide}
\title{Create Slide}
\usage{
Slide(rawText, number = NULL, totalNumber = NULL)
}
\arguments{
\item{rawText}{(character)}

\item{number, totalNumber}{(integer)}
}
\description{
Create Slide
}
