% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlideDeck.R
\name{SlideDeck}
\alias{SlideDeck}
\title{Create a slide deck}
\usage{
SlideDeck(slideSep = rep("", 3), newSlide = Slide)
}
\arguments{
\item{slideSep}{(character) what is used to indicate a slide break}

\item{newSlide}{(Slide) the 'Slide' constructor}
}
\description{
Create a slide deck
}
